move util to its own file
